split the retention anonymize act out of the dispatch table
The account arm took retentionactions.go past the 500-line ceiling. The split is by concept rather than by line count: the file is what the engine can DO to one over-age record, and anonymizing a person is the one action wide enough to be its own — the time-based twin of the Art. 17 cascade, held against it table by table.
